criação do factory do DataSplit
