+++ src/main/java/block/DataSplitFactoryImpl.java
@@ -1,0 +1,15 @@
+package block;
+
+public class DataSplitFactoryImpl extends DataSplitFactory{
+    DataSplit dataSplit = null;
+    @Override
+    public void createDataSplit(String logs, String data) {
+        dataSplit = new DataSplitImpl(logs,data);
+        dataSplit.insertData(data);
+    }
+
+    @Override
+    public void run(Double unitTime, Double initTime) {
+        dataSplit.run(unitTime,initTime);
+    }
+}
