+++ src/main/java/evaluate/Metrics.java
@@ -1,4 +1,4 @@
-package main.java.evaluate;
+package evaluate;
 
 import java.io.File;
